requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
sqlalchemy==1.4.46
pydantic==1.10.4
email-validator==1.3.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
psycopg2-binary==2.9.5
alembic==1.9.2
python-dotenv==0.21.1
websockets==10.4
aioredis==2.0.1
bcrypt==4.0.1